+++ src/components/shared/index.ts
@@ -1,0 +1,3 @@
+export { default as Button } from './Button';
+export { default as Dropdown } from './Dropdown';
+export { default as MultiSelect } from './MultiDropdown';
