+++ src/pages/test/index.tsx
@@ -1,6 +1,6 @@
-import Button from '@/components/Button';
-import Dropdown from '@/components/Dropdown';
-import MultiDropdown from '@/components/MultiDropdown';
+import Button from '@/components/shared/Button';
+import Dropdown from '@/components/shared/Dropdown';
+import MultiDropdown from '@/components/shared/MultiDropdown';
 import { useState } from 'react';
 
 const Test = () => {
@@ -13,7 +13,9 @@ const Test = () => {
         <Button variant="primary">버튼</Button>
         <Button variant="secondary">버튼</Button>
         <Button variant="outline">버튼</Button>
-        <Button variant="primary" disabled>버튼</Button>
+        <Button variant="primary" disabled>
+          버튼
+        </Button>
       </div>
 
       <div className="p-6">
